MOVEM is the first real instruction using the clocked CPU method. It is now a state machine, being called repeatedly during its execution until it reaches the finished state.
This means each word is actually read in 4 cycle intervals, and even
long writes will write in two separate bus accesses.

The test system has also been improved with some helpers, and a more
structured example of how instruction tests can be implemented.

The unobvious change to psg.c is because it needs to load SDL.h before
cpu.h is included (through mmu.h), because cpu.h now needs SDL_BYTEORDER.
This is because of how the read/write of MOVEM works.